+++ changelog.md
@@ -1,3 +1,7 @@
+### 2017-09-22
+[OptimShi]
+* Implemented "Actions" in the Movement Event (F74C)
+
 ### 2017-09-16
 [OptimShi]
 * Added PlayerDescriptionEvent 0x0013